chore: Rename .jsx file to .js. Apply clsx package.
