+++ package.json
@@ -2,11 +2,13 @@
   "name": "frontend",
   "version": "1.0.0",
   "private": true,
+  "proxy": "http://localhost:4001",
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.4",
+    "clsx": "^1.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.1",
