+++ src/App.test.js
@@ -1,5 +1,5 @@
 import { render, screen } from "@testing-library/react";
-import App from "./App.jsx";
+import App from "./App.js";
 
 test("renders learn react link", () => {
   render(<App />);
