+++ src/constants/index.js
@@ -1,5 +1,6 @@
 export const PageTitle = {
   HOME: "EGO Studying Online",
   SIGNIN: "EGO Studying Online | Signin",
-  SIGNUP: "EGO Studying Online | Signup"
+  SIGNUP: "EGO Studying Online | Signup",
+  NOTFOUND: "Page Not Found"
 };
