+++ src/pages/Home/Home.module.css
@@ -1,0 +1,7 @@
+.textAlignCenter {
+  text-align: center;
+}
+
+.textBold {
+  font-weight: bold;
+}
