+++ src/pages/index.js
@@ -1,3 +1,4 @@
 export { default as HomePage } from "./Home";
 export { default as SignupPage } from "./Signup";
 export { default as SigninPage } from "./Signin";
+export { default as NotFoundPage } from "./NotFound";
